Add to installers/workflows:  1) build *.zip medley-full release file for macos and 2) change all the release assets filename to follow the template:medley-full-<<PLATFORM>>-<<ARCH>>-<<MEDLEY.RELEASE>>_<<MAIKO.RELEASE>>

+++ installers/macos/build_app.sh
@@ -11,14 +11,23 @@
 
 APPNAME=Medley
 
+# 2023-07-19 Temorary selector to allow us to create .zip asset before app/dmg work is finished
+INCLUDE_APP="no"   # "yes" or "no"
+
 #  Make sure we are in the right directory
 if [ ! -f ./app/Info.plist ];
 then
-  echo "Can't find ./app/Info.plist file."
-  echo "Incorrect cwd?"
-  echo "Should be in medley/installers/macos"
-  echo "Exiting"
-  exit 1
+echo "Can't find ./app/Info.plist file."
+echo "Incorrect cwd?"
+echo "Should be in medley/installers/macos"
+echo "Exiting"
+exit 1
+fi
+
+# template for artifacts file names should be passed down in the ENV variable: ARTIFACTS_FILENAME_TEMPLATE
+if [ -z "${ARTIFACTS_FILENAME_TEMPLATE}" ];
+then
+  ARTIFACTS_FILENAME_TEMPLATE="medley-full-<<PLATFORM>>-<<ARCH>>-<<MEDLEY.RELEASE>>_<<MAIKO.RELEASE>>"
 fi
 
 #
@@ -26,11 +35,13 @@ fi
 #
 CWD=$(pwd)
 RESULTS_DIR=${CWD}/artifacts
-APPBUNDLE=${RESULTS_DIR}/${APPNAME}.app
-APPBUNDLECONTENTS=${APPBUNDLE}/Contents
-APPBUNDLEEXE=${APPBUNDLECONTENTS}/MacOS
-APPBUNDLERESOURCES=${APPBUNDLECONTENTS}/Resources
-APPBUNDLEICON=${APPBUNDLECONTENTS}/Resources
+if [ "${INCLUDE_APP}" = "yes" ]; then
+  APPBUNDLE=${RESULTS_DIR}/${APPNAME}.app
+  APPBUNDLECONTENTS=${APPBUNDLE}/Contents
+  APPBUNDLEEXE=${APPBUNDLECONTENTS}/MacOS
+  APPBUNDLERESOURCES=${APPBUNDLECONTENTS}/Resources
+  APPBUNDLEICON=${APPBUNDLECONTENTS}/Resources
+fi #INCLUDE_APP
 tmp_dir=${CWD}/tmp
 tarball_dir=${tmp_dir}/tarballs
 
@@ -78,47 +89,49 @@ fi
 pushd ${tarball_dir} >/dev/null 2>/dev/null
 medley_release=$(echo medley-*-loadups.tgz | sed "s/medley-\(.*\)-loadups.tgz/\1/")
 maiko_release=$(echo maiko-*-darwin.x86_64.tgz | sed "s/maiko-\(.*\)-darwin.x86_64.tgz/\1/")
-#debs_filename_base="medley-full-${medley_release}_${maiko_release}"
 popd >/dev/null 2>/dev/null
 
-#
-#  Create bundle dirs
-#
-rm -rf ${RESULTS_DIR}
-mkdir -p ${RESULTS_DIR}
-rm -rf ${APPBUNDLE}
-mkdir -p ${APPBUNDLE}
-mkdir ${APPBUNDLE}/Contents
-mkdir ${APPBUNDLE}/Contents/MacOS
-mkdir ${APPBUNDLE}/Contents/Resources
-#
-#  Create icons and put in bundle
-#
-iconset_dir=${tmp_dir}/${APPNAME}.iconset
-rm -rf ${iconset_dir}
-mkdir -p ${iconset_dir}
-image_dir=${CWD}/images
-sips -z 16 16     ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_16x16.png >/dev/null 2>&1
-sips -z 32 32     ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_16x16@2x.png >/dev/null 2>&1
-sips -z 32 32     ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_32x32.png >/dev/null 2>&1
-sips -z 64 64     ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_32x32@2x.png >/dev/null 2>&1
-sips -z 128 128   ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_128x128.png >/dev/null 2>&1
-sips -z 256 256   ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_128x128@2x.png >/dev/null 2>&1
-sips -z 256 256   ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_256x256.png >/dev/null 2>&1
-sips -z 512 512   ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_256x256@2x.png >/dev/null 2>&1
-sips -z 512 512   ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_512x512.png >/dev/null 2>&1
-cp ${image_dir}/App_icon1024.png ${iconset_dir}/icon_512x512@2x.png
-iconutil -c icns -o ${tmp_dir}/${APPNAME}.icns ${iconset_dir}
-cp ${tmp_dir}/${APPNAME}.icns ${APPBUNDLEICON}/
-rm -r ${iconset_dir}
-rm ${tmp_dir}/${APPNAME}.icns
-#
-#  Update and copy in "control" files
-#
-sed -e "s/--VERSION_TAG--/${medley_release}.${maiko_release}.0/g" \
-    < app/Info.plist \
-    > ${APPBUNDLECONTENTS}/Info.plist
-cp app/PkgInfo ${APPBUNDLECONTENTS}/
+if [ "${INCLUDE_APP}" = "yes" ]; then
+	#
+	#  Create bundle dirs
+	#
+	rm -rf ${RESULTS_DIR}
+	mkdir -p ${RESULTS_DIR}
+	rm -rf ${APPBUNDLE}
+	mkdir -p ${APPBUNDLE}
+	mkdir ${APPBUNDLE}/Contents
+	mkdir ${APPBUNDLE}/Contents/MacOS
+	mkdir ${APPBUNDLE}/Contents/Resources
+	#
+	#  Create icons and put in bundle
+	#
+	iconset_dir=${tmp_dir}/${APPNAME}.iconset
+	rm -rf ${iconset_dir}
+	mkdir -p ${iconset_dir}
+	image_dir=${CWD}/images
+	sips -z 16 16     ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_16x16.png >/dev/null 2>&1
+	sips -z 32 32     ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_16x16@2x.png >/dev/null 2>&1
+	sips -z 32 32     ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_32x32.png >/dev/null 2>&1
+	sips -z 64 64     ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_32x32@2x.png >/dev/null 2>&1
+	sips -z 128 128   ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_128x128.png >/dev/null 2>&1
+	sips -z 256 256   ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_128x128@2x.png >/dev/null 2>&1
+	sips -z 256 256   ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_256x256.png >/dev/null 2>&1
+	sips -z 512 512   ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_256x256@2x.png >/dev/null 2>&1
+	sips -z 512 512   ${image_dir}/App_icon1024.png --out ${iconset_dir}/icon_512x512.png >/dev/null 2>&1
+	cp ${image_dir}/App_icon1024.png ${iconset_dir}/icon_512x512@2x.png
+	iconutil -c icns -o ${tmp_dir}/${APPNAME}.icns ${iconset_dir}
+	cp ${tmp_dir}/${APPNAME}.icns ${APPBUNDLEICON}/
+	rm -r ${iconset_dir}
+	rm ${tmp_dir}/${APPNAME}.icns
+	#
+	#  Update and copy in "control" files
+	#
+	sed -e "s/--VERSION_TAG--/${medley_release}.${maiko_release}.0/g" \
+	    < app/Info.plist \
+	    > ${APPBUNDLECONTENTS}/Info.plist
+	cp app/PkgInfo ${APPBUNDLECONTENTS}/
+fi #INCLUDE_APP
+
 #
 #  Untar the maiko and medley releases into the bundle
 #
@@ -151,11 +164,9 @@ fileicon set ${il_dir}/medley/scripts/medley/medley.command ${image_dir}/Command
 #  Also create the zip file of il_dir for distribution
 #
 pushd ${il_dir} >/dev/null 2>&1
-zip -r -6 -y -q  ${RESULTS_DIR}/medley-full-${medley_release}_${maiko_release}-macos-universal.zip .
+filename="$(echo ${ARTIFACTS_FILENAME_TEMPLATE} | sed -e 's#<<PLATFORM>>#macos#' -e 's#<<ARCH>>#universal#' -e 's#<<MEDLEY.RELEASE>>#${medley_release}#' -e 's#<<MAIKO.RELEASE>>#${maiko_release}#' )"
+zip -r -6 -y -q  ${RESULTS_DIR}/${filename}.zip .
 popd >/dev/null 2>&1
 
-
-
-
-
+######################################################################################################
 ######################################################################################################
